==> scripts/Step4_analysis.Rmd <==
---
title: "Step3_preproc"
author: "zihan bai"
date: "2023-12-20"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

# Load necessary packages

```{r }
# install.packages("pacman)
library(pacman)
p_load(readr, dplyr, ggplot2, PupillometryR, PupilPre, shiny, tidyverse, stringr, emmeans, sjPlot, nlme, patchwork, purrr, ggpubr, lme4,PMCMRplus, stats, zoo, splines, psych)
```

# Paths

```{r}
# If work on Mac
base_path <- "/Users/bai/Library/CloudStorage/Box-Box/CAMSLab_Projects/Studies/AlcGen/Data/In_Lab"

# If work on Windows
# base_path <- "C:/Users/Zihan/Box/CAMSLab_Projects/Studies/AlcGen/Data/In_Lab"

# Which subjects?
subjects <- c('L003', 'L004', 'L006', 'L007', 'L009', 'L020', 'L021', 'H001', 'L022', 'L024', 'L026', 'L028', 'L029', 'L013', 'L014', 'L015', 'L017', 'L019', 'H002', 'H003', 'L023', 'L031')

excl_subjects <- c('L025')
```

# Conditioning

## CS onset

### Load preprocessed data

```{r}
# Load preprocessed data
preproc_path <- file.path(base_path, "pupillometry", "preprocessed")
preprocessed_cond_cs_path <- file.path(preproc_path, 'pupil_size', paste0('preprocessed_cond_cs.RData'))
if (file.exists(preprocessed_cond_cs_path)) {
  load(preprocessed_cond_cs_path)
}

# Check how many subjects are ready to go
n_distinct(subjects)
n_distinct(preprocessed_cond_cs$subject)
```

### Load behavior data

```{r}
# Load behavioral data
behav_fname <- paste0(base_path, "/", "processed_behav", "/", "combined_conditioning_data.csv")
behav_cond <- read_csv(behav_fname)
recode_behav_cond <- behav_cond %>%
  rename(trial = trial_num,
         subject = subId) %>%
  mutate(incong_loc = case_when(
          trial_cat %in% c("alc_null_surprise", "null_alc_surprise", 
                          "null_neut_surprise", "neut_null_surprise") ~ 'NOTHING',
          # ignore 30% trials while congruent locs contain nothing
          # so each cs is reinforced with the same us location in 70% of trials
          # and with the same us image in 85% of trials
          stim == 'null' ~ 'NOTHING',
          TRUE ~ as.character(incong_loc))) %>%
  filter(subject %in% subjects) %>%
  select(c(subject, phase, trial, incong_loc, stim, iter, drink_2grp))

n_distinct(recode_behav_cond$subject)
```

### Combine behavior and pupil
```{r}
combined_cond_cs <- left_join(preprocessed_cond_cs, recode_behav_cond, by = c("subject", "trial", "stim"))

combined_cond_cs$subject <- as.factor(combined_cond_cs$subject)
combined_cond_cs$time <- as.numeric(as.character(combined_cond_cs$time))

# Quick check to ensure row counts are as expected
nrow(combined_cond_cs)
nrow(preprocessed_cond_cs)
```

### Saccades and peak velocity during cs presentation
```{r}
# How do saccadic movements and their velocities differ in response to reward-related vs. non-reward stimuli, or alcohol vs non-alcohol?

saccade_cond_cs_path <- file.path(preproc_path, "pupil_size","saccade_cond_cs.RData")
if (file.exists(saccade_cond_cs_path)) {
  load(saccade_cond_cs_path)
}

subjects <- c('L003', 'L004', 'L006', 'L007', 'L009', 'L020', 'L021', 'H001', 'L022', 'L024', 'L026', 'L027', 'L028', 'L029')

sacc_behav <- behav_cond %>%
  filter(subject %in% subjects) %>%
  rename(trial = trial_num,
         subject = subId) %>%
  select(subject, trial, stim, iter, drink_2grp)

sacc_cond_cs <- saccade_cs_onset %>%
  bind_rows() %>%
  left_join(sacc_behav, by = c("subject", "trial"))

avg_sacc_metrics <- sacc_cond_cs %>%
  filter(subject != "H001") %>%
  group_by(subject, stim) %>%
  summarize(
    avg_sacc = mean(sacc_count, na.rm = TRUE),
    avg_mpv = mean(mpv, na.rm = TRUE),
    avg_max_pv = mean(max_pv, na.rm = TRUE),
    avg_min_pv = mean(min_pv, na.rm = TRUE),
    .groups = 'drop'
  )

ggplot(avg_sacc_metrics, aes(x = stim, y = avg_sacc, fill = stim)) +
  # geom_boxplot() +
  stat_summary(fun = mean, geom = "bar") +
  stat_summary(fun.data = mean_se, geom = "errorbar", width = 0) +
  labs(x = "CS stimulus",
       y = "Saccades") +
  scale_fill_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_minimal()
```


### Pupil response after cs onset by stimulus-type

```{r}
# For an aggregate view across subjects to highlight general trends rather than individual variations, we calcualte the mean size at each time point for each condition. This provides a representation of how the average pupil dilation changes over time in response to each stimulus type.

# Time-course
trace_plot_cond_cs <- combined_cond_cs %>%
  filter(stim %in% c("alc", "neut", "null")) %>%
  group_by(drink_2grp, time, stim) %>%
  summarise(avg_ps = mean(ps_cor, na.rm = TRUE), .groups = 'drop')

# check saccades frequency during window by stim

P_trace_plot_cs <- ggplot(trace_plot_cond_cs, aes(x = time, y = avg_ps, color = stim)) +
  facet_grid(~ drink_2grp) +
  geom_line() +
  labs(title = "Pupil dilation in response to CS presentation",
       x = "Time (s)",
       y = "Pupul dilation",
       color = "US") + 
  geom_vline(xintercept = 0.5, linetype = "dashed", color = "grey") +
  geom_vline(xintercept = 1.8, linetype = "dashed", color = "grey") +
  scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()
print(P_trace_plot_cs)

```

```{r}
ggplot(combined_cond_cs %>% mutate(cstype = case_when(stim %in% c("alc", "neut") ~ "CS+", stim == "null" ~ "CS-", TRUE ~ NA_character_)), aes(x = time, y = ps_cor, color = cstype)) +
  facet_grid(~ drink_2grp) +
  # geom_line() +
  geom_smooth(se=FALSE) +
  scale_color_manual(values = c("CS+" = "#A06593", "CS-" = "grey")) +
  theme_minimal()
```

```{r}
# Save the figure
pdf(file = 'P_trace_plot_cs.pdf', width = 8, height = 4, useDingbats = F)
print(P_trace_plot_cs)
dev.off()
```

### Pre/Post incongruent pupil response

#### Method 2

```{r}
# Understand how pupil size changes around trials where the US location is incongruent with the usual reinforcement location.

# Select distinct rows for relevant columns
dat_per_stim <- combined_cond_cs %>%
  select(subject, stim, trial, incong_loc) %>%
  distinct()

# Recode order per stimulus type
dat_per_stim <- dat_per_stim %>%
  group_by(subject, stim) %>%
  mutate(stimrep = row_number())

# Preparing incongruent trials data frames
incong_loc_trials <- dat_per_stim %>%
  filter(incong_loc == "TRUE") %>%
  mutate(prerep = stimrep - 1, postrep = stimrep + 1)

incong_outcome_trials <- dat_per_stim %>%
  filter(incong_loc == "NOTHING") %>%
  mutate(prerep = stimrep - 1, postrep = stimrep + 1)

# Initialize flags
dat_per_stim$pre_incong_loc <- 0
dat_per_stim$post_incong_loc <- 0
dat_per_stim$pre_incong_outcome <- 0
dat_per_stim$post_incong_outcome <- 0

# Check each trial to see if it directly precedes or follows an incongruent trial for the same subject and stimulus. If it does, flag it by 1
for (i in 1:nrow(dat_per_stim)) {
  current_row <- dat_per_stim[i, ]
  # Check for pre and post incongruent location
  if (any(incong_loc_trials$prerep == current_row$stimrep & incong_loc_trials$subject == current_row$subject & incong_loc_trials$stim == current_row$stim)) {
    dat_per_stim$pre_incong_loc[i] <- 1
  }
  if (any(incong_loc_trials$postrep == current_row$stimrep & incong_loc_trials$subject == current_row$subject & incong_loc_trials$stim == current_row$stim)) {
    dat_per_stim$post_incong_loc[i] <- 1
  }
  # Check for pre and post incongruent outcome
  if (any(incong_outcome_trials$prerep == current_row$stimrep & incong_outcome_trials$subject == current_row$subject & incong_outcome_trials$stim == current_row$stim)) {
    dat_per_stim$pre_incong_outcome[i] <- 1
  }
  if (any(incong_outcome_trials$postrep == current_row$stimrep & incong_outcome_trials$subject == current_row$subject & incong_outcome_trials$stim == current_row$stim)) {
    dat_per_stim$post_incong_outcome[i] <- 1
  }
}

pre_post_incong_loc <- combined_cond_cs %>%
  left_join(dat_per_stim, by = c("subject", "stim", "trial", "incong_loc")) %>%
  mutate(incong_loc_timing = case_when(
    pre_incong_loc == 1 ~ "pre",
    post_incong_loc == 1 ~ "post",
    TRUE ~ as.character(NA))) %>%
  filter(!is.na(incong_loc_timing))

# Plotting
ggplot(pre_post_incong_loc, aes(x = time, y = ps_cor, color = incong_loc_timing)) +
  facet_grid(~ stim~drink_2grp) +
  geom_smooth(se = TRUE) +
  labs(x = "Time(s)", y = "Pupil Dilation", color = "Trial", title = "Pupil dilation in response to CS presentation", subtitle = "Pre/post incongruent locations") +
  theme_minimal()

pre_post_incong_out <- combined_cond_cs %>%
  left_join(dat_per_stim, by = c("subject", "stim", "trial", "incong_loc")) %>%
  mutate(incong_loc_timing = case_when(
    pre_incong_outcome == 1 ~ "pre",
    post_incong_outcome == 1 ~ "post",
    TRUE ~ as.character(NA))) %>%
  filter(!is.na(incong_loc_timing), stim != "null")

# omit double-counted trials
# table(pre_post_incong_loc[pre_post_incong_loc$subject=='L006',]$trial, rowSums(cbind(pre_post_incong_loc[pre_post_incong_loc$subject=='L006',]$pre_incong_loc, pre_post_incong_loc[pre_post_incong_loc$subject=='L006',]$post_incong_loc))) << if anything is in the '2' column, it is double-counted

# Plotting
ggplot(pre_post_incong_out, aes(x = time, y = ps_cor, color = incong_loc_timing)) +
  facet_grid(~ stim~drink_2grp) +
  geom_smooth(se = TRUE) +
  labs(x = "Time(s)", y = "Pupil Dilation", color = "Trial", title = "Pupil dilation in response to CS presentation", subtitle = "Pre/post incongruent outcomes") +
  theme_minimal()
```

#### Method 1

```{r}
# Compare pre vs post incongruent trials
# unique instances of trials for each subject and stimulus type, focusing on whether the trial was incongruent or not
dat_per_stim <- combined_cond_cs %>% select(c("subject", "stim", "trial", "incong_loc")) %>% distinct()

# Recode order per stimulus type
dat_per_stim <- dat_per_stim %>% group_by(subject, stim) %>% mutate(stimrep = row_number())

# Flag pre/post incong. Calculate indices for the trial immediately preceding and following an incongruent trial or null trial
incong_loc_trials <- dat_per_stim %>% filter(incong_loc=="TRUE") %>%
  mutate(prerep = stimrep-1,
         postrep = stimrep+1)

incong_outcome_trials <- dat_per_stim %>% filter(incong_loc=="NOTHING") %>%
  mutate(prerep = stimrep-1,
         postrep = stimrep+1)

# Add info to dat_per_stim df
for (s in sort(unique(dat_per_stim$subject))) {
  dat_sub <- dat_per_stim %>% filter(subject==s)
  incong_loc_sub <- incong_loc_trials %>% filter(subject==s)
  incong_outcome_sub <- incong_outcome_trials %>% filter(subject==s)
  
  dat_sub$pre_incong_loc = 0
  dat_sub$post_incong_loc = 0
  
  dat_sub$pre_incong_outcome = 0
  dat_sub$post_incong_outcome = 0
  
  
  for (stimname in sort(unique(dat_per_stim$stim))) {
    dat_stim <- dat_sub %>% filter(stim==stimname)
    # location of US
    stim_pre_idx <- which(dat_stim$stimrep %in% incong_loc_sub[incong_loc_sub$stim==stimname,]$prerep)
    dat_stim[stim_pre_idx,]$pre_incong_loc = 1
    
    stim_post_idx <- which(dat_stim$stimrep %in% incong_loc_sub[incong_loc_sub$stim==stimname,]$postrep)
    dat_stim[stim_post_idx,]$post_incong_loc = 1
    
    # US type
    stim_pre_idx <- which(dat_stim$stimrep %in% incong_outcome_sub[incong_outcome_sub$stim==stimname,]$prerep)
    dat_stim[stim_pre_idx,]$pre_incong_outcome = 1
    
    stim_post_idx <- which(dat_stim$stimrep %in% incong_outcome_sub[incong_outcome_sub$stim==stimname,]$postrep)
    dat_stim[stim_post_idx,]$post_incong_outcome = 1
    
    if (stimname=='alc') {
    dat_stim_concat <- dat_stim
    } else {
      data_stim_concat <- rbind(dat_stim_concat, dat_stim)
    }
  } # stim loop
  
  if (s=='H001') {
    dat_sub_concat <- data_stim_concat
  } else {
    dat_sub_concat <- rbind(dat_sub_concat, data_stim_concat)
  }
} # subject loop

# plot pre- and post (where are the other stim...)
ggplot(combined_cond_cs %>% merge(dat_sub_concat, by = c("subject", "stim", "trial", "incong_loc")) %>%
         mutate(incong_loc_timing = ifelse(pre_incong_loc==1, "pre", ifelse(post_incong_loc==1, "post", NA))) %>% filter(!(is.na(incong_loc_timing))), 
       aes(x = time, y = ps_cor, color = incong_loc_timing)) +
  facet_grid( stim~ drink_2grp) +
  # geom_line() +
  geom_smooth(se=FALSE) +
  labs(title = "Pupil dilation in response to CS presentation: Pre/post incongruent locations",
       x = "Time (s)",
       y = "Pupul dilation",
       color = "US") + 
  geom_vline(xintercept = 0.5, linetype = "dashed", color = "grey") +
  geom_vline(xintercept = 1.8, linetype = "dashed", color = "grey") +
  # scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()

ggplot(combined_cond_cs %>% merge(dat_sub_concat, by = c("subject", "stim", "trial", "incong_loc")) %>%
         mutate(incong_outcome_timing = ifelse(pre_incong_outcome==1, "pre", ifelse(post_incong_outcome==1, "post", NA))) %>% filter(!(is.na(incong_outcome_timing))), 
       aes(x = time, y = ps_cor, color = incong_outcome_timing)) +
  facet_grid( stim~ drink_2grp) +
  # geom_line() +
  geom_smooth(se=FALSE) +
  labs(title = "Pupil dilation in response to CS presentation: Pre/post incongruent outcomes",
       x = "Time (s)",
       y = "Pupul dilation",
       color = "US") + 
  geom_vline(xintercept = 0.5, linetype = "dashed", color = "grey") +
  geom_vline(xintercept = 1.8, linetype = "dashed", color = "grey") +
  # scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()
```

### Window analysis

```{r}
# In contrast, we calculate the mean pupil changes for each condition for each subject and controls for the individual differences among subjects
start_window <- 0.5
end_window <- 1.8
# end_window <- 1

win_cond_cs <- combined_cond_cs %>%
  filter(time >= start_window & time <= end_window,
         stim %in% c("alc", "neut", "null")) %>%
  group_by(subject, stim, drink_2grp) %>%
  summarize(win_avg = mean(ps_cor, na.rm = TRUE), .groups = 'drop')

P_win_cs <- ggplot(win_cond_cs, aes(x = stim, y = win_avg, fill = stim)) +
  facet_grid(~ drink_2grp) +
  geom_boxplot() +
  labs(x = "US",
       y = "Mean pupil dilation",
       fill = "US") + 
  scale_fill_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()
print(P_win_cs)
```

```{r}
# Save the figure
pdf(file = 'P_win_cs.pdf', width = 4, height = 4, useDingbats = F)
print(P_win_cs)
dev.off()
```

### Within-subject pupil changes by condition

```{r}
subject_condition_averages <- win_cond_cs %>%
  group_by(drink_2grp, subject, stim) %>%
  summarise(mean_ps_cor = mean(win_avg, na.rm = TRUE), .groups = 'drop')

P_win_ind_cs <- ggplot(subject_condition_averages, aes(x = stim, y = mean_ps_cor, group = subject)) +
  facet_grid(~ drink_2grp) +
  geom_line(aes(color = subject), show.legend = FALSE) + # Connects points with lines for each subject
  geom_point(aes(color = subject)) + # Plots the actual points
  labs(title = "Within-subject pupil size across conditions",
       x = "CS",
       y = "Mean pupil size") +
  theme_minimal() +
  scale_color_viridis_d() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
print(P_win_ind_cs)
```

```{r}
# Save the figure
pdf(file = 'P_win_ind_cs.pdf', width = 4, height = 4, useDingbats = F)
print(P_win_ind_cs)
dev.off()
```

### Split into halves

```{r}
group_avg_cond_cs_split <- combined_cond_cs %>%
  mutate(split = case_when(iter<=20 ~ "firstH", iter>20 ~ "secondH")) %>%
  group_by(split, time, stim, drink_2grp) %>%
  summarise(avg_ps = mean(ps_cor, na.rm = TRUE),
            se = sd(ps_cor, na.rm = TRUE) / sqrt(n()),
            .groups = 'drop')
P_trace_plot_cs_split <- ggplot(group_avg_cond_cs_split, aes(x = time, y = avg_ps, color = stim, group = interaction(stim, drink_2grp))) +
  geom_line() +
  facet_grid(drink_2grp ~ split) + # Adjusted for a grid by drinker group and split
  labs(title = "Pupil dilation in response to CS presentation",
       x = "Time (s)",
       y = "Pupil dilation", # Corrected typo
       color = "Stimulus Type") +
  geom_vline(xintercept = 0.5, linetype = "dashed", color = "grey") +
  geom_vline(xintercept = 1.8, linetype = "dashed", color = "grey") +
  scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()
print(P_trace_plot_cs_split)
```

```{r}
# Save the figure
pdf(file = 'P_trace_plot_cs_split.pdf', width = 4, height = 4, useDingbats = F)
print(P_trace_plot_cs_split)
dev.off()
```

```{r}
win_cond_cs_split <- combined_cond_cs %>%
  filter(time >= start_window & time <= end_window) %>%
  mutate(split = case_when(iter<=20 ~ "firstH", iter>20 ~ "secondH")) %>%
  group_by(split,subject, stim, drink_2grp) %>%
  summarize(win_avg = mean(ps_cor, na.rm = TRUE), .groups = 'drop')

P_win_cond_cs_split <- ggplot(win_cond_cs_split, aes(x = stim, y = win_avg, fill = split)) +
  facet_grid(~ drink_2grp) +
  geom_boxplot() +
  labs(x = "US",
       y = "Pupil dilation",
       color = "US") + 
  theme_classic()
print(P_win_cond_cs_split)
```

```{r}
# Save the figure
pdf(file = 'P_win_cond_cs_split.pdf', width = 4, height = 4, useDingbats = F)
print(P_win_cond_cs_split)
dev.off()
```

### Statistical Analysis

```{r}
# the effects of stimulus type, drinking level, and repetition on pupil dilation.
model_aov <- aov(mean_ps ~ stim, data = group_avg_cond_cs)
summary(model_aov)

# Model pupil size as a function of stimulus type, drinking level, and repetition, including random effects for subject
model <- lmer(ps ~ stim + (1 | subject), data = combined_cond_cs)

summary(model)
```

### Repetition

```{r}
# divide into iterations
cond_cs_iter <- combined_cond_cs %>%
  filter(stim %in% c("alc", "neut", "null")) %>%
  group_by(subject, stim, iter) %>%
  summarize(avg_ps = mean(ps_cor, na.rm = TRUE)) %>%
  ungroup()

# lm
lm_cond_cs_iter <- lme(avg_ps ~ stim * iter, random=~1|subject, data = cond_cs_iter)
anova_results <- anova(lm_cond_cs)
print(anova_results)
  
plot_model(lm_cond_cs, type = "eff", terms = c("iter", "stim"), axis.title = c("Repetition", "Pupil dilation")) +
  scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5", "null" = "grey")) +
  theme_classic()
```

## Towardness ~ cond

```{r}
# Which subjects?
subjects <- c('L003', 'L004', 'L006', 'L007', 'L009', 'L020', 'L021', 'H001', 'L022', 'L024', 'L026', 'L027', 'L028', 'L029')

excl_subjects <- c('L025')

# In what sample rate?
s_rate_500 <- c('L003', 'L004', 'L006', 'L007', 'L009', 'L020', 'L021', 'H001')
s_rate_250 <- c('L013', 'L014', 'L015', 'L017', 'L019', 'H002', 'H003')
s_rate_1000 <- c('L025', 'L031')

# Define paths
base_path <- "/Users/bai/Library/CloudStorage/Box-Box/CAMSLab_Projects/Studies/AlcGen/Data/In_Lab"
```

```{r}
# Load anticipation data
append_cond_gaze_path <- file.path(preproc_path, "gaze", "append_cond_gaze.RData")
append_cond_gaze <- list()
if (file.exists(append_cond_gaze_path)) {
  load(append_cond_gaze_path)
}

for (subject in subjects) {
  input_path <- file.path(trial_split_path, subject, paste0(subject, "_cond_split.RData"))
  data <- load(input_path)
  ant <- subject_phases_list$ant_onset %>%
    filter(!is.na(incong_loc)) %>%
    rename(trial = trial_num) %>%
    group_by(trial) %>%
    mutate(time = (row_number() - 1) / 1000 * 2) %>%
    ungroup()
  
  append_cond_gaze[[subject]] <- ant
}
save(append_cond_gaze, file = append_cond_gaze_path)

append_cond_gaze_df <- bind_rows(append_cond_gaze)

for (subject in subjects) {
  input_path <- file.path(trial_split_path, subject, paste0(subject, "_cond_split.RData"))
  data <- load(input_path)
  fix <- subject_phases_list$fixation %>%
    filter(!is.na(incong_loc)) %>%
    rename(trial = trial_num) %>%
    group_by(trial) %>%
    mutate(time = (row_number() - 1) / 1000 * 2) %>%
    ungroup()
  
  append_cond_fix[[subject]] <- fix
}
append_cond_fix_df <- bind_rows(append_cond_fix)
```

```{r}
gaussian_kernel <- function(sd, len = 30) {
  seq <- seq(-len/2, len/2, length.out = len)
  kernel <- dnorm(seq, mean = 0, sd = sd)
  kernel / sum(kernel)
}
sd_points <- 5  # Standard deviation in terms of data points
kernel <- gaussian_kernel(sd = sd_points)
```

```{r}
# Before averaging across participants, gaze time courses were smoothed by a Gaussian kernel with a 10 ms s.d. (Ede et al., 2019)

# Linear Interpolation on the raw data to fill in missing values.
# Gaussian Kernel Smoothing of the interpolated data to reduce noise.
# Normalization of the smoothed data using a baseline period to control for individual differences.
# Calculation of towardness from the smoothed and normalized data.

Sdata <- make_pupillometryr_data(data = append_cond_gaze_df, subject = subject, trial = trial, time = time, condition = stim)

missing <- calculate_missing_data(Sdata, pupil = xp)
excl_data <- clean_missing_data(Sdata,
                                pupil = xp,
                                trial_threshold = .75,
                                subject_trial_threshold =.75)
int_data <- interpolate_data(data = excl_data, pupil = xp, type = "linear")
  

smooth_data <- int_data %>%
  group_by(subject, trial) %>%
  mutate(xp = ifelse(is.na(xp), NA, 
                                          stats::filter(xp, kernel, sides = 2))) %>%
  ungroup()

fix_median <- append_cond_fix_df %>%
  group_by(subject, trial) %>%
  summarize(fix_median = median(xp, na.rm = TRUE)) %>%
  ungroup()

fix_corr <- smooth_data %>%
  left_join(fix_median, by = c("subject", "trial"))
  
fix_norm <- fix_corr %>%
  group_by(subject, trial) %>%
    # figure out expected position of US
  mutate(expect_loc = ifelse(stim == "null", "none",
                               ifelse(incong_loc == "FALSE", us_loc,
                                      ifelse(us_loc=="right", "left",
                                             ifelse(us_loc=="left", "right", NA))))) %>%
  mutate(xp_norm.fix = (xp - fix_median) / fix_median, # positive = right, negative = left
         towardness.fix = ifelse(expect_loc=="left", -1 * xp_norm.fix, 
                               ifelse(expect_loc=="right", xp_norm.fix,
                                           ifelse(expect_loc == "none", NA, NA)))) %>%
  ungroup()

# smooth_fix_norm <- fix_norm %>%
#   group_by(subject, trial) %>%
#   mutate(smoothed_towardness.fix = ifelse(is.na(towardness.fix), NA, 
#                                           stats::filter(towardness.fix, kernel, sides = 2))) %>%
#   ungroup()
         
```

```{r}
preprocessed_towardness <- file.path(preproc_path, "gaze", "preprocessed_cond_towardness.RData")
load(preprocessed_towardness)

P_cond_towardness <- ggplot(fix_norm %>% filter(stim != "null"), aes(x = time, y = towardness.fix, color = stim)) + geom_smooth(se = TRUE) + labs(y="Towardness", x="Time") + scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5")) + theme_classic() + geom_hline(yintercept = 0)

P_cond_sidebias <- ggplot(fix_norm %>% filter(stim == "null"), aes(x = time, y = xp_norm.fix, color = stim)) + geom_smooth(se = FALSE) + labs(y="Towardness", x="Time") + scale_color_manual(values = c("null" = "grey")) + theme_classic() + geom_hline(yintercept = 0)

```

```{r}
# Save the figure
pdf(file = 'P_cond_towardness.pdf', width = 4, height = 4, useDingbats = F)
print(P_cond_towardness)
dev.off()

pdf(file = 'P_cond_sidebias.pdf', width = 4, height = 4, useDingbats = F)
print(P_cond_sidebias)
dev.off()

# Save the file
output_towardness_path <- file.path(base_path, "pupillometry", "preprocessed", "gaze", "preprocessed_cond_towardness.RData")
save(fix_norm, file = output_towardness_path)
```

```{r}
# look for reaction time for alc and neut because the time they take to move to the expected location seems different?
```

## US onset (Secondary)

```{r}

```

### Pupil response after us onset by stimulus-type

```{r}
# Time-course
ggplot(merged_us_df, aes(x = time, y = ps_cor, group = stim, color = stim)) +
  stat_summary(fun = mean, geom = "line", na.rm = TRUE) +
  theme_minimal()

# Window analyses
# Window of interest (0.5-1.8s)
window_start = 0.5
window_end = 1.8
combined_us_window <- filter(merged_us_df, between(time, window_start, window_end))
us_window <- create_window_data(combined_us_window, pupil = ps_cor)
plot(us_window, pupil = ps_cor, windows = F, geom = 'boxplot')

```

# Generalization

## GS onset

### Load behavior and pupil

```{r}
preprocessed_gen_gs_path <- file.path(preproc_path, 'pupil_size', paste0('preprocessed_gen_gs_test.RData'))
if (file.exists(preprocessed_gen_gs_path)) {
  load(preprocessed_gen_gs_path)
}

subjects <- unique(combined_data$subject)

# Load behavior in generalization task
behav_fname <- paste0(base_path, "/", "processed_behav", "/", "combined_generalization_data.csv")

behav_gen <- read_csv(behav_fname) %>%
  filter(block != 'prac') %>%
  rename(subject = subId,
         diss = gs_deg,
         stim = trial_cat) %>%
  group_by(subject) %>%
  mutate(trial = row_number()) %>%
  ungroup() %>%
  filter(subject %in% subjects) %>%
  select(subject, phase, trial, stim, diss, drink_2grp)
```

### Combine behavior and pupil

```{r}
combined_gen_gs <- left_join(combined_data, behav_gen, by = c("subject", "trial", "stim"))

combined_gen_gs$subject <- as.factor(combined_gen_gs$subject)
combined_gen_gs$time <- as.numeric(as.character(combined_gen_gs$time))
# Quick check to ensure row counts are as expected
nrow(combined_gen_gs)
nrow(combined_data)
```

### Pupil response after gs onset by stimulus-type

```{r}
# Card-evoked pupil dilation will be analyzed with a linear mixed effects model with degree of similarity to original CS, US category, drinking level, and age as predictors, with time spent doing the task as a covariate of no interest.
P_traceplot_gen_gs <- ggplot(combined_gen_gs %>% filter(stim!="novel"), aes(x=time, y=ps_cor, color=as.factor(diss))) + 
  geom_smooth() + 
  facet_grid(~ stim~drink_2grp, scales = "free_y") + 
  labs(x = "Time", y = "Pupil dilation", color = "Dissimilarity") +
  theme_minimal()
print(P_traceplot_gen_gs)

ggplot(combined_gen_gs %>% filter(stim!="novel"), aes(x=time, y=ps_cor, color=stim)) +
     geom_smooth() + 
     facet_grid(~ as.factor(diss), scales = "free_y") + 
     labs(x = "Time", y = "Pupil dilation", color = "Dissimilarity") +
     theme_minimal()
```

```{r}
# Save the figure
pdf(file = 'P_traceplot_gen_gs.pdf', width = 8, height = 4, useDingbats = F)
print(P_traceplot_gen_gs)
dev.off()
```

## Towardness ~ gen

```{r}
# two plots, each contains diss for one stimulus
# Which subjects?
subjects <- c('L003', 'L004', 'L006', 'L007', 'L009', 'L020', 'L021', 'H001', 'L022', 'L024', 'L026', 'L027', 'L028', 'L029')

# Define paths
base_path <- "/Users/bai/Library/CloudStorage/Box-Box/CAMSLab_Projects/Studies/AlcGen/Data/In_Lab"

# Input
trial_split_path <- file.path(base_path, "pupillometry", "restructured", "trial_split")

# Output
preproc_path <- file.path(base_path, "pupillometry", "preprocessed")
if (!dir.exists(preproc_path)) {
  dir.create(preproc_path)
}
```

```{r}
# Load anticipation data
append_gen_ant_path <- file.path(preproc_path, "gaze", "append_gen_ant.RData")
append_gen_ant <- list()
if (file.exists(append_gen_ant_path)) {
  load(append_gen_ant_path)
}

for (subject in subjects) {
  input_path <- file.path(trial_split_path, subject, paste0(subject, "_gen_split.RData"))
  data <- load(input_path)
  ant <- subject_phases_list$ant_onset %>%
    rename(trial = trial_num,
           stim = trial_cat,
           diss = gs_deg) %>%
    group_by(trial) %>%
    mutate(time = (row_number() - 1) / 1000 * 2) %>%
    ungroup() %>%
    select(subject, time, xp, yp, ps, trial, stim, diss, drink_2grp)
    
  append_gen_ant[[subject]] <- ant
}
save(append_gen_ant, file = append_gen_ant_path)

# USE THIS: gaze during anticiaption

append_gen_ant_df <- bind_rows(append_gen_ant)

# Load fixation data
append_gen_fix_path <- file.path(preproc_path, "gaze", "append_gen_fix.RData")
append_gen_fix <- list()
if (file.exists(append_gen_fix_path)) {
  load(append_gen_fix_path)
}
for (subject in subjects) {
  input_path <- file.path(trial_split_path, subject, paste0(subject, "_gen_split.RData"))
  data <- load(input_path)
  fix <- subject_phases_list$fixation %>%
    rename(trial = trial_num,
           stim = trial_cat,
           diss = gs_deg) %>%
    group_by(trial) %>%
    mutate(time = (row_number() - 1) / 1000 * 2) %>%
    ungroup() %>%
    select(subject, time, xp, yp, ps, trial, stim, diss, drink_2grp)
  
  append_gen_fix[[subject]] <- fix
}
save(append_gen_fix, file = append_gen_fix_path)

# USE THIS: gaze during fixation
append_gen_fix_df <- bind_rows(append_gen_fix)
```

```{r}
gaussian_kernel <- function(sd, len = 30) {
  seq <- seq(-len/2, len/2, length.out = len)
  kernel <- dnorm(seq, mean = 0, sd = sd)
  kernel / sum(kernel)
}
sd_points <- 5  # Standard deviation in terms of data points
kernel <- gaussian_kernel(sd = sd_points)
```

```{r}
# Before averaging across participants, gaze time courses were smoothed by a Gaussian kernel with a 10 ms s.d. (Ede et al., 2019)

# Linear Interpolation on the raw data to fill in missing values.
# Gaussian Kernel Smoothing of the interpolated data to reduce noise.
# Normalization of the smoothed data using a baseline period to control for individual differences.
# Calculation of towardness from the smoothed and normalized data.

Sdata <- make_pupillometryr_data(data = append_gen_ant_df, subject = subject, trial = trial, time = time, condition = stim)

missing <- calculate_missing_data(Sdata, pupil = xp)
excl_data <- clean_missing_data(Sdata,
                                pupil = xp,
                                trial_threshold = .75,
                                subject_trial_threshold =.75)
int_data <- interpolate_data(data = excl_data, pupil = xp, type = "linear")

smooth_data <- int_data %>%
  group_by(subject, trial) %>%
  mutate(xp = ifelse(is.na(xp), NA, 
                                          stats::filter(xp, kernel, sides = 2))) %>%
  ungroup()

fix_median <- append_gen_fix_df %>%
  group_by(subject, trial) %>%
  summarize(fix_median = median(xp, na.rm = TRUE)) %>%
  ungroup()

fix_corr <- smooth_data %>%
  left_join(fix_median, by = c("subject", "trial"))

# Find the expect location for each subject and stimulus during gen task based on cond task
behav_fname <- paste0(base_path, "/", "processed_behav", "/", "combined_conditioning_data.csv")
behav_cond <- read_csv(behav_fname)
recode_behav_cond <- behav_cond %>%
  rename(trial = trial_num,
         subject = subId) %>%
  filter(subject %in% subjects) %>%
  group_by(subject, stim) %>%
  summarize(expect_loc = case_when(
    incong_loc == FALSE & stim == "neut" ~ us_loc,
    incong_loc == FALSE & stim == "alc" ~ us_loc,
    TRUE ~ NA_character_
  ), .groups = 'drop') %>%
  distinct() %>%
  na.omit()

combined_gen_ant <- fix_corr %>%
  left_join(recode_behav_cond, by = c("subject", "stim")) %>%
  mutate(expect_loc = case_when(
    stim == "novel" ~ "none",
    stim == "null" ~ "none",
    TRUE ~ expect_loc
  ))
  
fix_norm <- combined_gen_ant %>%
  group_by(subject, trial) %>%
    # figure out expected position of US
  mutate(expect_loc = case_when(stim == "null" ~ "none",
                                stim == "novel" ~ "none", TRUE ~ expect_loc)) %>%
  mutate(xp_norm.fix = (xp - fix_median) / fix_median, # positive = right, negative = left
         towardness.fix = ifelse(expect_loc=="left", -1 * xp_norm.fix, 
                               ifelse(expect_loc=="right", xp_norm.fix,
                                           ifelse(expect_loc == "none", NA, NA)))) %>%
  ungroup()
```

```{r}
preprocessed_towardness <- file.path(preproc_path, "gaze", "preprocessed_gen_towardness.RData")
load(preprocessed_towardness)

P_gen_towardness <- ggplot(fix_norm %>% filter(stim %in% c("alc", "neut")), aes(x = time, y = towardness.fix, color = stim)) + geom_smooth(se = TRUE) + labs(y="Towardness", x="Time") + scale_color_manual(values = c("alc" = "#A06593", "neut" = "#6086C5")) + theme_classic() + geom_hline(yintercept = 0)

P_gen_sidebias <- ggplot(fix_norm %>% filter(stim %in% c("null", "novel")), aes(x = time, y = xp_norm.fix, color = stim)) + geom_smooth(se = TRUE) + labs(y="Towardness", x="Time") + scale_color_manual(values = c("null" = "grey", "novel" = "grey23")) + theme_classic() + geom_hline(yintercept = 0)

```

```{r}
preprocessed_gen_gaze_path <- file.path(preproc_path, "gaze", "preprocessed_gen_gaze.RData")
if (file.exists(preprocessed_gen_gs_path)) {
  load(preprocessed_gen_gs_path)
}
         
```



```{r}
# Save the figure
pdf(file = 'P_gen_towardness.pdf', width = 4, height = 4, useDingbats = F)
print(P_gen_towardness)
dev.off()

pdf(file = 'P_cond_sidebias.pdf', width = 4, height = 4, useDingbats = F)
print(P_cond_sidebias)
dev.off()

# Save the file
output_towardness_path <- file.path(base_path, "pupillometry", "preprocessed", "gaze", "preprocessed_cond_towardness.RData")
save(fix_norm, file = output_towardness_path)
```
# Behavior
```{r}
home_dir <- '/Users/bai/Library/CloudStorage/Box-Box/'
# home_dir <- '~/Library/CloudStorage/Box-Box/'
proj_dir <- paste0(home_dir, 'CAMSLab_Projects/Studies/AlcGen/')
data_dir <- paste0(proj_dir, 'Data/In_Lab/processed_behav/') # update to include more subj later
conditioning_concat_export <- read.csv(paste0(data_dir, 'combined_conditioning_data.csv'))
gen_concat_export <- read.csv(paste0(data_dir, 'combined_generalization_data.csv'))
lucky_concat_export <- read.csv(paste0(data_dir, 'combined_rating_data.csv'))
combined_usmem_data <- read.csv(paste0(data_dir, 'combined_usmem_data.csv'))
combined_csmem_data <- read.csv(paste0(data_dir, 'combined_csmem_data.csv'))

# redefine variables for later use
conditioning_concat <- conditioning_concat_export
lucky_concat <- lucky_concat_export
gen_concat <- gen_concat_export
usmem_concat <- combined_usmem_data
csmem_concat <- combined_csmem_data
#mem_concat <- mem_concat_export
#survey_dat <- survey_dat_export %>% mutate(education = as.numeric(education))

# so far, manually adding ages
age_df <- data.frame(subId = c('L003', 'L004', 'L005', 'L006', 'L007',  'L009',  'L011', 'L012', 'L013', 'L014', 'L015',
                               'H001', 'H002',  'H003'),
                     age = c(23, 36, 47, 69, 76, 54, 41, 26, 37, 22, 27,
                             67, 41, 68)
                     )
# would need to export regularly from RedCap to ensure subId and age list are updated, but saved in box under analysis_live


# define luckiness data (conditioning)
cs_img_info <- conditioning_concat %>% filter(stim %in% c("alc", "neut", "null")) %>%
  select(c("drink_2grp", "subId", "stim", "cs_img")) %>% distinct()
cs_img_info <- cs_img_info %>%
  mutate(cs_img_name = ifelse(startsWith(str_sub(cs_img_info$cs_img,-11), "V") ,str_sub(cs_img_info$cs_img,-11),str_sub(cs_img_info$cs_img,-10)))

lucky_concat <- lucky_concat %>%
rename(cs_img_name = stim)

lucky_concat$drink_2grp <- NA
lucky_concat[grepl("L",lucky_concat$subId),]$drink_2grp <- "light"
lucky_concat[grepl("H",lucky_concat$subId),]$drink_2grp <- "heavy"
lucky_compare <- merge(lucky_concat, cs_img_info, by = c("subId", "cs_img_name","drink_2grp")) %>%
  mutate(timing = factor(timing, levels = c("pre", "post")),
         response = as.numeric(response))

img_luck_mod <- lme(response ~ timing*stim*drink_2grp, random=~1|subId, 
                    dat = lucky_compare %>%
                      mutate(timing = factor(timing, levels = c("pre", "post")),
                             stim = factor(stim, levels = c("null", "alc", "neut"))), na.action = na.omit)

car::Anova(img_luck_mod, type=2, test.statistic = "Wald")

# posthoc
emmeans(img_luck_mod, pairwise ~ timing|stim)$contrasts 
# emmeans(img_luck_mod, pairwise ~ drink_2grp|timing|stim)$contrasts

lucky_compare %>%
  mutate(timing = factor(timing, levels = c("pre", "post")),
         stim = factor(stim, levels = c("null", "alc", "neut"))) %>%
  filter(subId %in% subjects) %>%
  ggplot(aes(x = timing, y = response, fill = timing)) +
  stat_summary(fun = mean, geom = "bar", position = "dodge", alpha = .8) +
  stat_summary(fun.data = mean_se, geom = "errorbar", position = position_dodge(0.9), width = 0) +
  geom_point(position = position_dodge(0.9), alpha = .6, aes(color = timing), size = .5) +
  geom_line(aes(group = subId), alpha = .4) +
  facet_grid(.~stim) +
  theme_classic()
```
```{r}
# define choice data (generalization)
gen_final <- gen_concat %>% filter(!(block=="prac")) %>%
  mutate(stim = ifelse(grepl("alc", trial_cat), "alc",
                       ifelse(grepl("neut", trial_cat), "neut",
                              ifelse(grepl("null", trial_cat), "null",
                                     ifelse(grepl("novel", trial_cat), "novel", NA))))) %>%
  merge(cs_img_info %>% select(-c("drink_2grp")),
        by = c("subId", "stim"), all.x = TRUE) %>%
  mutate(play_bin = ifelse(gs_choice == -1, 0, 
                           ifelse(gs_choice == 1, 1, NA)),
         deg_plot = ifelse(gs_deg == 999, 0, gs_deg),
         deg_scale = sjmisc::center(as.numeric(deg_plot)),
         stim = factor(stim, levels = c('null', 'novel', 'neut', 'alc')))
gen_final$drink_2grp <- ""
gen_final[grepl("L",gen_final$subId),]$drink_2grp <- "light"
gen_final[grepl("H",gen_final$subId),]$drink_2grp <- "heavy"

# define RT data (generalization)
probe_rt_dat <- gen_concat %>%
  filter((trial_cat != "novel") & (block != "prac") & (probe_acc == 1)) %>%
  mutate(gs_rt_log = log(gs_rt),
         probe_rt_log = log(probe_rt)) %>%
  mutate(deg_plot = ifelse(gs_deg == 999, 0, gs_deg),
         trial_cat = factor(trial_cat, levels = c("neut","null", "alc"))) %>%
  filter_all(any_vars(!is.na(.)))

# choice generalization
gen_final <- gen_final %>%
  merge(age_df, by = "subId") %>%
         mutate(drink_2grp = ifelse(grepl("L", subId), "light", "aud"))

binmod_2grp_reward <- glm(play_bin ~ trial_cat*deg_plot*drink_2grp + age + (trial_num+1), 
                          dat = gen_final %>% filter(!(trial_cat  %in% c("novel", "prac"))) %>%
                            mutate(play_bin = ifelse(gs_choice == -1, 0, 
                                                     ifelse(gs_choice == 1, 1, NA)),
                                   deg_plot = ifelse(gs_deg == 999, 0, gs_deg), 
                                   trial_cat = factor(trial_cat),
                                   subId = factor(subId)),
                          family = "binomial")

summary(binmod_2grp_reward)
car::Anova(binmod_2grp_reward, test.statistic = "Wald")
plot_model(binmod_2grp_reward, type = "eff", terms = c("deg_plot", "trial_cat", "drink_2grp"))
```
